Add validation checks for stock and price in inventory methods

+++ domain/use_cases/inventory_use_cases.py
@@ -10,9 +10,9 @@ class InventoryUseCases:
             raise ValueError("Product ID is required")
         if not inventory.name:
             raise ValueError("Name is required")
-        if not inventory.stock:
+        if not inventory.stock or inventory.stock != 0:
             raise ValueError("Stock is required")
-        if not inventory.price:
+        if not inventory.price or inventory.price != 0:
             raise ValueError("Price is required")
         if self.inventory_repo.get_inventory(inventory.product_id):
             raise ValueError("Inventory already exists")
@@ -21,6 +21,8 @@ class InventoryUseCases:
                                                     inventory.description)
 
     def get_inventory(self, product_id: int) -> Inventory:
+        if not self.inventory_repo.get_inventory(product_id):
+            raise ValueError("Inventory does not exist")
         return self.inventory_repo.get_inventory(product_id)
 
     def update_inventory(self, product_id: int, inventoryUpdate: InventoryUpdate) -> InventoryUpdate:
@@ -29,4 +31,6 @@ class InventoryUseCases:
         return self.inventory_repo.update_inventory(product_id, inventoryUpdate)
 
     def delete_inventory(self, product_id: int) -> bool:
+        if not self.inventory_repo.get_inventory(product_id):
+            raise ValueError("Inventory does not exist")
         return self.inventory_repo.delete_inventory(product_id)
